+++ ribs/archives/_elite.py
@@ -20,7 +20,7 @@ class Elite(NamedTuple):
     #: Measure values.
     measures: np.ndarray
 
-    #: Index of the elite in the archive (see :meth:`ArchiveBase.get_index`).
+    #: Index of the elite in the archive (see :meth:`ArchiveBase.index_of`).
     index: int
 
     #: Metadata object for the elite.