+++ ribs/archives/_grid_archive.py
@@ -105,35 +105,48 @@ class GridArchive(ArchiveBase):
 
     @staticmethod
     @jit(nopython=True)
-    def _get_index_numba(behavior_values, upper_bounds, lower_bounds,
-                         interval_size, dims):
-        """Numba helper for get_index().
+    def _index_of_numba(behavior_values, upper_bounds, lower_bounds,
+                        interval_size, dims):
+        """Numba helper for index_of().
 
-        See get_index() for usage.
+        See index_of() for usage.
         """
         # Adding epsilon to behavior values accounts for floating point
         # precision errors from transforming behavior values. Subtracting
         # epsilon from upper bounds makes sure we do not have indices outside
         # the grid.
-        behavior_values = np.minimum(
-            np.maximum(behavior_values + _EPSILON, lower_bounds),
-            upper_bounds - _EPSILON)
+        measures = np.minimum(np.maximum(measures + _EPSILON, lower_bounds),
+                              upper_bounds - _EPSILON)
 
-        index = (behavior_values - lower_bounds) / interval_size * dims
-        return index.astype(np.int32)
+        indices = (measures - lower_bounds) / interval_size * dims
+        return indices.astype(np.int32)
 
     def index_of(self, measures):
+        """Returns indices of the behavior values within the archive's grid.
+        First, values are clipped to the bounds of the behavior space. Then, the
+        values are mapped to cells; e.g. cell 5 along dimension 0 and cell 3
+        along dimension 1.
+        The indices can be used to access boundaries of a behavior value's cell.
+        For example, the following retrieves the lower and upper bounds of the
+        cell along dimension 0::
+            idx = archive.index_of(...)  # Other methods also return indices.
+            lower = archive.boundaries[0][idx[0]]
+            upper = archive.boundaries[0][idx[0] + 1]
+        See :attr:`boundaries` for more info.
+        Args:
+            measures (numpy.ndarray): (:attr:`behavior_dim`, batch_size) array of
+                coordinates in behavior space.
+        Returns:
+            tuple of int: The grid indices.
+        """
         # Adding epsilon to behavior values accounts for floating point
         # precision errors from transforming behavior values. Subtracting
         # epsilon from upper bounds makes sure we do not have indices outside
         # the grid.
-        measures = np.minimum(
-            np.maximum(measures + _EPSILON, self._lower_bounds),
-            self.upper_bounds - _EPSILON)
-
-        indices = (measures -
-                   self._lower_bounds) / self._interval_size * self._dims
-        return np.ravel_multi_index(indices.astype(np.int32), self._dims)
+        indices = self._index_of_numba(behavior_values, self._upper_bounds,
+                                       self._lower_bounds, self._interval_size,
+                                       self._dims)
+        return self.ravel_index(indices)
 
     # TODO: Docstrings.
     def ravel_index(self, index):